Pagination refactoring
The admin/video/index action now uses the tg.decorators.paginate helper
instead of handling pagination directly. We still uses webhelpers.paginate
under the hood but is much more convenient. Unfortunately it doesn't apply
easily transfer to the dashboard, so for now I've left that action alone.

show-more.js has been refactored. Heights are no longer hardcoded in
px, URLs just modify the 'page' query string param and leave the rest
intact, and if no URL is provided the current window location is used.
(This means that when we're using the tg paginate deco we don't need to
bother specifying the current URL as the ajax endpoint.) Finally I've
added a 'onMove' event which fires before after the newly loaded content
is injected into the page but before it is revealed.