Frontend templating changes + Mediaflow matches new mockup + Mootools 1.2.2
Boxes have a new div structure with less verbose naming and it should be
more flexible as well. <box> tags are replaced with all the necessary
<div> tags as before, but now it does this in place without altering the
rest of the document.

For example:
	<box id="asdf">
		<nav>
			optional HTML will be appended to the default nav
		</nav>
		box content goes here
	</box>

Results basically in:
	<div id="asdf" class="box">
		<div class="box-top" />
		<div class="box-nav">default nav here, optional HTML appended</div>
		<div class="box-body">box content goes here</div>
		<div class="box-bottom" />
	</div>

I may unify section specific nav in the future, still fleshing that out.
Also, the current nav will change because podcasts isn't ready yet.

Finally, we've already been using Mootools 1.2.2 on the backend for a
while now, now we're using it on the frontend too.