Refactor comments admin, still some lingering issues to fix
The way EditCommentForm's are being generated is problematic since it
currently generates a form for ALL comments in the database, not just
the comments on the given page. Its quite possible that the comment data
and the comment forms won't match up either, so you could easily be
overwriting one comments text with another comments text.

This approach seems to go against the grain of tw.forms IMHO. It will be
cleaner and better on performance to create one form, with one save
action, which we can reuse thusly:

	<div py:for="c in comments" py:content="edit_form(c)" />

This means we'd have to update the form to accept model instances as
values, which seems logical to me anyway.

This does mean our access logs would not show which comment is being
approved/deleted etc since the URLs wouldn't have the ID in them (it'd
have to be passed in a hidden input. I'm willing to accept this
downside unless anyone has a better idea.

See mediaplex/controllers/commentadmin.py for the FIXME notes.