Fix comment edit form creation w.r.t. pagination
Simply substituted the list comprehension in the controller for a
generator expression so that it isn't evaluated until we iterate over it
in the template, at which point pagination has taken place.

I still feel like we should be reusing the same form instance for all
these forms, passing in a different comment for rendering each step of
the way. See the commit message on cb7396575 'Refactor comments admin'.