Fix admin search form overtext and value display
The overtext code was commented out -- I can only assume this was done
for development purposes since it seems to work flawlessly. Looks like
mel you commented it out in an unrelated commit.

Also the value was never being displayed because the value was being
passed as a kwarg, not a dict:
	search_form(search=search)  !=  search_form(dict(search=search))