Add deployment database dump.
Can be used instead of the devel.sql dump for testing purposes.
I've deleted as many spam comments as I could find, and normalized all of the URLs and the media types.

Dump was created with the following command:

mysqldump5 -u root mediaplex | perl -p -e s:),(:),n(:g > deploy.sql

I think we should use this command for all database dumps: The key is that the perl replace script creates linebreaks on each row, so we can tell via git which precise rows are updated in a particular revision of a database dump.