Fix markup/unicode problems with templates
the Markup() call was just a temporary fix for a data import from old blogs
and the str() call was causing unicode decode errors, because it was converting unicode data to byte strings.
it should never have been there. It was introduced in february in the root.html template, and just never removed.