Update config + thumb sizes + podcast album art uploader
I've changed the config.media_thumb_dir to config.image_dir so that we
don't need to create new config options for every module's thumb dir.

I've made a AlbumArtUploader class which should work for most use cases.
It uses my Uploader class which is just a wrapper of Swiff.Uploader. It
has an option for updating form actions when you upload a file to some
new content which doesn't yet exist in the database -- the ID is passed
back in the JSON and all the form actions in the page are updated.

Fixes all the thumb sizes for media and podcasts on the frontend. Also,
podcast album art is now square 1:1 in accordance with the iTunes spec.