Cleanup podcast style
Moved the podcast series info to the top of the podcast/view page.

Reduced the description length on the podcast episode grid. Typically
the description should span 2 lines and then the height will line up
perfectly vertically, but if it will not cause any clipping problems
if it spans 3 lines. The only downside is if the description isn't
long enough, but that shouldn't be happening often, if ever.

Also corrected a few XHTML properties which weren't being wrapped with
Markup(x).stripentities()