Disable autocomplete on admin forms in firefox.
The issues we were seeing previously with incorrect form values
after a page refresh are actually due to a firefox bug. Argh.
This just adds a domready event to disable autocomplete on all forms
in the admin page. We should, perhaps, add exceptions to this rule,
for convenience when searching.