Fix uploader.
There was an error when no tags were submitted; the validator would complain
that the tags key was missing (swiff uploader doesn't submit empty fields).

I've added an if_missing value to correct this.

Also, there was an issue where the upload() method was validating its own form.
This caused the form fields to all be marked as errors by default. Oops!