Fix logic errors in truncate_xhtml helper
XHTML was not stripped, even if the option was enabled, if the initial string
was less than 'size' characters long. Oops!

Also, extra processing was being done, if _strip_xhtml was enabled, the Cleaner() was
being run needlessly, and if _strip_xhtml was disabled, the block_spaces replacement regex
was being run needlessly.