Rewrite the template fetching code in base.RoutingController
Now it's compatible with multiple concurrent processes.
It uses a lock file (php_new.html) to handle concurrency.
And uses the mtime of the template file, rather than an instance variable
to handle timeouts.

Only works on unix systems, because file renames are guaranteed to work when
a file with that name already exists, and are also guaranteed to be atomic.