Fix file ordering wrt to adding new files -- DB updated
New files are now added added at a position 1 greater than the last,
starting at 0.

Because of the way the Media.reposition_file algo works, it's possible
that we could've eventually hit the upper bound of the
media_files.position database column (UNSIGNED TINYINT is 255) so
I've changed the column type to UNSIGNED SMALLINT which maxes out at 65k
which should be more than enough. Also, because of how the media_files
SQLAlchemy Table is defined, I moved the column to the bottom/right.

Finally, I was getting that KeyError content_type issue while using
Request.JSON so I switched to using the plain Mootools Request and
JSONifying the response manually. Not sure what's going on here really..