Correctly serve akamai files as linked to by podcasts
Podcasts always output links such as this:
http://localhost:8080/media-files/who-do-you-think-you-are-2366.mp3

Our on-site media_player links to Akamai files directly, but for
podcasts, we want a consistent URL structure, so MediaController.serve
needs to redirect for us. This patch adds that behaviour.

Consistent URLs are good because they make Feedburner stats more legible
and RSS readers will never get confused if we move the file from our
server to a CDN. And we'd likely to run into that soon, when we start
using a cron job to move uploaded files to Akamai, rather than
forcing the user to wait for it to happen before their upload is
considered "complete".