Fix uploading via swiff.uploader on windows by changing the returned content-type.
This is not the standard way to deliver JSON. It's supposed to be "application/json".

Looks like certain combinations of Flash / Windows send an accept header that
only likes "text/*" for responses to file uploads. This is an issue when you
try to return JSON as content-type "application/json". TG will choke instead
of sending an unacceptable content-type.

There is no way to override the header in the Swiff.Uploader, though, because
it uses the FileReference.upload() method. Which, for some reason, doesn't allow
custom headers.