Presenting squeezebox to notify user that comment is pending approval
When a user submits a comment and it is held for moderation there is no action that lets them know it is held. This leads to users re-submitting comments because they thought they didn't go through when in fact they did. We just need a simple javascript notifier to let users know that their comment has been received and is pending moderator approval.