Genericize lib.base.BaseController: external template now configurable
I've removed a lot of TG-quickstart crap that we weren't using. All
controllers now inherit the lib.base.BaseController, which inherits
RoutingController. If someone were to want to use TGController for its
object dispatch, the change would only have to be made in one place.

I've created simpleplex.templates.layout. This should as a deploy-
specific means of wrapping the box of content in a layout matching
wherever the site is deployed to. The name of this layout config can
be changed in the INI file, so you have the option of copying the
file so Git won't consider editing it a change to your working tree.

I've also genericized the external PHP template loader. There are now
INI config options for:
	external_template
		Flag to enable or disable use of the external template
	external_template_name
		The name to load/save the external template as.
	external_template_url
		The URL to pull the external template from
	external_template_timeout
		The number of seconds before the template should be refreshed

Now it's easy to have the external config used in one deployment and not
in another... which is handy for development.