Fix header handling when serving media files
If a client requests 'text/*' we throw a 406 Not Acceptable, as stated
in the HTTP specification.

The response Content-Type used to have an erroneous ';charset=utf8'
appended to it, that has been removed.

Content-Disposition has been added, so when files are directly requested
they download instead of play in browser (or show gibberish for FLV
video, for example).