Correct duration overlay markup/naming in media/podcast grids
Block-level elements should not be used inside inline elements. I'm
pretty sure it's not allowed in XHTML Strict, but not only that, but if
you display the content without CSS the display will be more retarded
than it needs to be.

This *MAY* fix an intermittent rendering error. Either that or it's
Genshi getting seriously confused under some circumstance and producing
bogus markup.

Couple other things:
- the class naming did not much the names of its related grid classes,
  I added the grid- prefix and changed 'timestamp' to 'duration'.
- the padding on the left side of the duration was greater than
  on the right so I evened it out.