Fix exposing custom content types with picky Accept headers, closes #2
This problem goes deeper in TG than I initially thought. We needed
to be able to serve JSON as 'text/plain' when given an Accept request
header that is simply 'text/*' -- but normally we still want to serve
the correct 'application/json' header. My solution is to use tg's
CUSTOM_CONTENT_TYPE flag and manually set the header I wanted depending
on whether the Accept header allowed for. That's all fine and dandy but
CUSTOM_CONTENT_TYPE doesn't work when the Accept header doesn't contain
'*/*'. This patch works around that added problem by extending our
base controller class.

There's a proper fix for this in the tg 2.1 branch, which is
still in beta. If we find the time we will backport it.