Add default_comment_status config option, skip approval by default
A setting of 'unreviewed' means the comment must be approved by a
moderator before the comment is published on the site. In this case, the
user gets a squeezebox notification after the form is submitted to say
so.

Setting 'publish', the default, means the comment goes straight up on
the site.