Cleanup admin CSS by replacing .mo with .btn
Everywhere a button was used, so too was 'mo'. There was a lot of
duplication and inconsistency with how button classes were named, where
they were placed, and how they were used, so this fixes it. Just use the
.btn base class, add .btn-my-button with background-image, width and
height. Rollover and everything is taken care of.

This some makes generic reusable styles out of duplicated styles/images.

The categories and tags interfaces are still broken. I'll be fixing them
in subsequent commits.