Separate categories and tags controllers/forms from other settings
Although our navigation bar groups categories, tags, and users under
'Settings', this was done due to the limitation of the design -- not
because they are conceptually 'Settings'. It is more important for the
underlying module structure to be conceptually sound than it is for it
to match the current nav/URL scheme, which is likely to change anyway.