Update frontend to use the configured Featured Category whenever possible
If there is no featured category, or no media that matches the given
filter criteria, then the first popular item is used instead. This means
that an item will be display as being featured any time there is media.

If the Featured Category setting is set to a non-existant ID,
it's the same as setting it to None. It doesn't break anything.