Update MediaFile and everything that depends on it, UI revision coming
We're assuming now that all files associated with a given media object
will contain essentially the same content. Multiple videos associated
with one media item are simply encoded with different settings (for
different devices and qualities), not entirely different videos.

Captions and audio descriptions can only be added to videos. If there is
both audio and video for a single media item, the audio file is assumed
to be an audio description, and the videos are the primary content.