Fix URL generation for /admin/settings/categories,tags,users
Without map.explicit = False, the fallback route looks like the best
match, so that's what resolves, despite the order the routes are
defined in. This is a change in behaviour between Routes 1.11 and 1.12.

We should consider removing the fallback route, as we don't actually
need it with our current routes, and it creates ambiguity in URLs.