Improve filetypes module and how the Media.encoded flag is set
The Media.encoded flag is now set by trying to pick a file to play,
rather than looking for our arbitrarily blessed file extensions.
Podcast media is only encoded if there is a file that plays in iTunes,
regular media is only encoded if there is a file that plays in Chrome.
These choices will become configurable in a later release.