Improve continuity when uploading a file/thumb before saving media
The title and slug used as the name of the media stub are put into the
form, and you can update the publish status right away. Also, the 'New
Media' heading is replaced with the new media title, which is a link to
the actual edit page for the newly created media.