Resolve simultaneous meta file/thumb/meta saving by merging after each save
If you save the main meta data form while the first file on a new media
item is uploading, both these requests will create their own media item
in the database. However, after the second (or third) request completes,
we 'merge' it into the first item that completed.

With this commit, adding and editing media is 100% ajax. The only
downside to this is that the page never refreshes to /admin/media/ID, it
stays on /admin/media/new until you navigate back to the page, or click
the heading at the top of the page (which becomes a link after you save
the first time).