Fix lots of issues with add/edit media ajax
Saving the main meta form now updates the status form:
Putting media in a podcast may cause it to be set not-encoded.

Update the files table with new XHTML after merging stubs, as well as
the status form.

Update the page title and box heading when changing the title of the
media.

Improves the order of operations when deciding to merge media.

Fixes race condition that occurred when uploading files with Safari 5:
onSelectStart would create file.ui too slowly, and onUploadStart would
already be trying to use it.