Add a monkey patch to avoid unnecessary redirects.
StaticURLParser is enabled when static_files is set to true in the
config file.

By default, it likes to issue a 301 redirect to cannonicalize any
requests for public folder names if they are missing a trailing slash.
(eg. /scripts would be redirected to /scripts/)

This caused problems with routes that mimicked folder names.
(eg /admin would be redirected to /admin/)

Since MediaCore doesn't use public directory listings, or index.html
files inside the public directory, I've removed that default behaviour
with this patch, and thus avoided unnecessary HTTP requests.