Enforce alphabetical sort order on playable media files.
Sort criteria for playable media files is now:
- Video before Audio
- File Size (descending)
- Container name (ascending)

This allows us to guarantee which file will be played, given a known
combination of preferred player, media files, and browser.

It also rearranges the logic so that all of the get_*_player()
subfunctions have the same overall "for file in..." format.