Conform to the HTML5/Safari spec for audio and video tags
While the HTML5 spec includes a "poster" attribute that displays an
image of our choosing before the video plays, it unfortunately flips to
the first frame of the video as soon as it loads. Its still better than
nothing, it'll be seen on some mobile devices like the iPhone where
autobuffering is always off.