Handle concurrency exceptions when incrementing media views
Don't raise an exception should concurrency problems occur.
Views will not actually be incremented in this case, but thats
relatively unimportant compared to rendering the page for the user.

We may be able to remove this after we improve our triggers to not
issue an UPDATE on media_fulltext unless one of its columns are
actually changed. Even when just media.views is updated, all the
columns in the corresponding media_fulltext row are updated, and
media_fulltext's MyISAM engine must lock the whole table to do so.