Fix positioning of submit button on upload form.
This is a fairly dirty hack to fix a problem inherent to the HTML
structure of the form. While a table seemed like a good idea for our
initial upload form requirements, a List would now be more appropriate.

You will see that the submit buttom drifts downwards still, if a file
with a long filename is chosen. This is considered acceptable for now,
as a complete rewrite of the form template is beyond the scope of this
point release.