Add missing licenses, rename all external license files.
Some third party packages have licences included inline.
For all packages where this is not the case, an external
file should be added.
This file will have the word LICENSE in capital letters
as part of the file name, and will either name the
licensed package, or reside in the subdirectory of the
licensed package.