Prepare all form text for babel i18n message extraction
The i18n that happens in a class definition happens on application
startup. Even though we pass a lazy gettext function, which I would
expect to be evaluated when the widget is rendered, it is instead
evaluated when the class is loaded. This is unacceptable in the
long run since we want to be able to localize dynamically, based on
where the user is coming from, not one localization per instance.

Instead we must internationalize when rendering inside Genshi. By
wrapping all text in N_() we ensure that `setup.py extract_messages`
will find these messages and the localization of the string can be
done, and then later found at render time.