Use relative paths for all template lookups
This paves the way for making @expose usable in plugins, where we have
multiple loaders in place. The old approach of always using an absolute
template path to invoke absolute path resolution in <xi:include> is no
longer suitable. We want the Genshi template loader to decide which
template to load, not our expose decorator. Furthermore, by using
explicit relative paths in <xi:include> we avoid collisions with
plugins.