Add initial plugin loading architecture
We use setuptools entry points to find plugins that have been installed
to the virtualenv that MediaCore runs in. These plugins can currently
define their own templates, which can be included on top of core
templates if they follow the appropriate naming convention; static files
that will be served by the paste StaticURLParser app; and controllers.