Add basic events which plugins can subscribe to
The core design philosophy here is that these events should not be
coupled to the application code that calls them. This allows us to avoid
import-time complications.

Our forms widgets are instantiated when controllers are imported, and
then reused for the life of the application. If events were attached
directly to the controller, plugin authors would have to be careful to
import the form and attach a post_init hook BEFORE importing the
controller which instantiates it, which is likely to be a common use
case. It would be great to move to ToscaWidgets2 at some point, which
is designed to have widgets instantiated once per request. We may also
want to consider deform, though I haven't looked at it very closely yet.