Extend Media feed support
Add a latest media feed and link to explore and media and make some
minor fixes to the xml to ensure validate and compatible.

Add category feed per category (order by date) and add the links for this
feed to the category pages. Render latest media feed when no category
selected