Fix migration 015: Create MultiSettings
For this script to work with paster setup-app, we must call Table.create
and .drop with (checkfirst=True). This is because setup-app creates all
tables before running any migrations, so we essentially try to create
this table twice. We can probably improve on this logic, I copied it
from an example posted to the SQLAlchemy list by Michael Bayer.