Improve setup-app logic for running migrations
Once a database has been initialized with a migrate version, the only
changes that will be made to it will be done by sqlalchemy-migrate.

Before now, we were using a technique outlined by Michael Bayer, the
lead developer of SQLAlchemy. When a new table is created, setup-app
would create the table BEFORE running the migrate-upgrade command, so
the migrate script that was intended to create the table effectively
does nothing. This isn't likely to cause problems in many situations, so
long as you say create your table with checkfirst=True, but this
solution is safer.