Initial Import of selenium tests
A simple suite and basic test layout for mediacore selenium
testing.  Test the frontend and some backend admin features.

Create 2 main suites one for frontand and another for backend and a
fullsuite that runs all the tests from both main suites.
Add tests for the rest of the admin section

Includes test for bulk media delete