Render URIs using StorageURI.__str__ for its autoconcatenation
With this change it becomes possible to build a StorageURI with a file
name and a base URL and have it automatically concatenated upon render.
This means that when we're building URIs, we don't have to special case
RTMP urls -- when the player receives a StorageURI with the rtmp scheme,
it alone knows what special treatment the filename and server uri require.