Add paste gzip middleware and enable it by default
From the sounds of things, this doesn't actually add a statistically
significant drain on the CPU under normal loads. While it's better to
enable gzip in Apache, enabling it here is second best. We'll just have
to document that this can be disabled if you're going through the
trouble of enabling it in Apache.