Fix admin comments page.
Includes new delete icon, to be used across all admin table forms.

Reworks the publish/delete form to actually be a form, using POST requests.
This is made necessary by the changes introduced in
56d4c0dbd7e2a27677da6701c4e0896fa3edc9cb

Also reworks the javascript AJAX managers for the comments page, enabling
the javascript to work with the new document structure and clarifying some
old logic.