Always return '' for gettext('')
This comes up in the context of genshi.filters.i18n where <img> alt
attributes are provided for XHTML validation, but are empty because
the only applicable alt text would be redundant. The genshi filter
translates the empty string and the result is the headers at the top of
the mediacore.po file.