Flesh out JW Embedder externs to enable proper type checking
The externs we had before worked because the compiler isn't too
aggressive with renaming externs, even in advanced mode.  Without
enabling a switch that's currently only available in the programmatic
java API, for any X, X.setup() is not renamed because jwplayer.setup is
a known extern.  In other words methods aren't renamed event if their type
should indicate that it's safe unless we explicitly opt into it -- which
I'd like very much to do at some point.