Improve Translator performance
This will reuse the same translator for all requests until the primary
language changes. This should reduce file system access because, even
though the gettext module caches message catalogs, it still does a bunch
of file system access to "find" the correct message catalog before it
realizes it has a cached copy.

Additionally in a speed comparison, I found that this technique:
	return getattr(msgid, 'domain', None)

is approximately 2-3x slower than this other technique when the domain
attribute isn't set (which is the most common):
	if isinstance(msgid, TranslateableUnicode): return msgid.domain