Improve playerbar popups to autoselect the permalink and open new windows
The auto focus works for the embed code as well.

I'm using a somewhat hack-ish technique of adding an invalid
'target="_blank"' attribute on anchors to open a new window. This
attribute isn't valid XHTML 1.0 Strict but it's a lot easier than adding
event listeners for each link to open the window with javascript.