Improve JWPlayer expand/shrink on pageload
When the page is rendered with the widescreen setting enabled, but the
user's preference is the smaller size, the resize is now more seamless.
We now resize the flash <object> tag while it's loading, then call the
proper JW Embedder resize API once the player has finished loading.
Prior to this the container around the <object> was resized, but the
object still had a fixed (larger) size.