Fix char encoding issue by removing unnecessary XML() calls. Closes #62
We used to need to use XML() but now it seems that it's no longer needed
and indeed is unnecessary overhead. This may be necessary when ToscaWidgets
uses its own Genshi template loader, but in v0.9 TW templates are loaded
using the same loader as all other templates, so it seems the streams are
now compatible.

XML() casts its argument to string and then back to another stream. It's
during this cast that the character encoding issue was rearing its head,
in the Russian translation on the categories and tags pages. By working
with a continuous markup stream the whole way through we apparently avoid
the character encoding issue entirely.