Comment out the additional notes option, pending its complete removal
If this setting is enabled it will still work -- and no data will be
lost, but we're hiding the settings for it and disabling it by default.
If no one complains we can safely assume that no one is using it and it
will be removed in v0.9.1, or supported with a plugin.