Disable automatic HTML entity decoding in normal text fields
While this can be handy at times, it prevents certain inputs from being
entered properly. Ampersands are commonly stripped if they are not
surrounded by whitespace, which ensures that they not appear to be
HTML entities. We're just going to have to trust our users to use
characters not entities...