Remove transparency on cooliris background
With the alpha set to 0, the background was transpsarent and causing a
very ugly green screen to be shown when Cool Iris was in full screen
mode. By changing that, we can allow people to set the background color
of their Cool Iris wall and also retain that color when full screen mode
is engaged.