Display the locale ID (eg. es_VE) in the language select field
This also ends the practice of consistently leaving the primary_language
setting empty to mean the default English translation. We still
interpret the empty setting to mean English, but when the form is saved,
we explicitly store English as the preference. This fixes a stupid
sorting issue that forced us to put the default translation as the first
option, away from the other English translations.