Revert "Add setting to control tiling of background image"
This temporarily reverts commit e8906b3b224d144581904205ca436a8cc7846077.

This is a feature addition that belongs in 0.9.1, not this release.
Note that there is a bug whereby you can save the value of this setting,
but the radio list is not populated with the value when you re-visit the page.