Replace access to app_globals.settings with request.settings
This sets us up well for porting to pyramid, since the request object is
passed around explicitly and holds nearly all the functionality of the
current pylons stacked object proxies.

The settings object is added to the template globals for performance,
since we do a lot of settings-related stuff in our templates.