Fix all media_count_published subquery column properties
Prior to this, any time the query is evaluated, the application
startup time would be used as the cut off for the publish date.

This is the first place we are using on the database time, every other
datetime is explicitly specified by our python code. If the database
server time or tz is misconfigured, this could cause some weird issues.