use HTML5 based markup for Facebook comments unless the xid was stored in the media's meta data.
Facebook will display a scary user-visible warning if an XFBML/xid based markup is used and
there are no comments stored ("Warning: this comments plugin is operating in compatibility
mode, but has no posts yet.") so we default to the newer HTML5/href based markup.

Add a command-line batch script to query Facebook and store the xid permanently if
comments exist for a media. New installs (or media added after the upgrade) will always
use the href based markup.

There are two shortcomings in the new approach which result in the loss of comments:
 - changing the domain name (or MediaCore reachable under different domains)
 - manually changing the slug
However both issues are fixable with an approach similar to the xid/href decision
implement in this commit.