ability to detect db connection leakage (by default disabled in deployment.ini).
"Leakage" refers to DB connections which were retrieved from the DB pool but not returned by the end of the request. This results in higher resource usage but leaked connections are also not subject to MediaCore's session cleanup. This means the code using a leaked connection will never recover from a bad connection (e.g. OperationalError, server went away) until the process is restarted.

Connection leakage is always a programmer error and should be treated as a bug.