update MooTools core to latest version their 1.2 stable branch (1.2.6dev, git revision a3be7a0ef5205eb655c4ee11e8cbedcb3f2ffa95).
The update makes the MediaCore admin functional again in Firefox 18. Before we had broken squeezeboxes (e.g. "Add & Manage Files") as well as several layout issues.

MooTools caused the bug because they modify String.prototype in a way that clashes with EcmaScript 6 and Firefox 18 was obviously the first browser to implement that part of ES6.

More information can be found in the MooTools issue 2402 ("MooTool's String.prototype.contains conflicts with ES6"):
  https://github.com/mootools/mootools-core/issues/2402