remove compiled translation files ('.mo') for all locales where less than 40% of all strings are translated.
Languages with so many missing translations likely don't provide a good user experience but I don't want to purge them because there was still some valuable work which I'd like to preserve.

Without a .mo file a language will not appear in the backend language selection but it can be enabled by compiling the .mo-file manually.

The following languages are therefore disabled by default:
- Armenian/Armenia ("hy_AM")
- Dutch ("nl")
- Greek ("el")
- Macedonian ("mk")
- Spanish/Costa Rica ("es_CR")
- Tamil ("ta)