Plugins can now ship a "db_default.py" module with an "add_default_data()" function to modify the database in case the plugin does not have any db state stored.
MediaDrop will assume that after successfully executing the method the
db schema/contents will be the same as if the latest alembic migration script was
executed.

To enable this functionality plugins must also have alembic migrations set up
(even if they don't have any revisions). Each plugin should be prepared for DB
migrations if they introduce custom structures.

This setup is better than listing to events because it integrates with the
alembic flow: MediaDrop will create a migration table and stamp it with the
currently highest revsion automatically. Even the case of "no migrations but default
data" is handled.