ignore controller methods which do not allow GET requests in redirect after login (came_from)
This avoids a potentially confusing 405 ("method not allowed") error. For example
an admin user will be logged out after changing his password but the login
page gets a "came_from" parameter pointing to /admin/users/XY/save. So if the admin
just logs in again he would be "greeted" with the error message above.